requirements.txt
torch>=1.7.0
torchvision>=0.8.0
opencv-python>=4.5.0
numpy>=1.19.0
pandas>=1.2.0
scikit-learn>=0.24.0
tqdm>=4.60.0
Pillow>=8.0.0
scipy>=1.6.0

